backend/interests/requirements.txt
Django>=2.0,<3.0
django-cors-headers
djangorestframework
djangorestframework_simplejwt
django-debug-toolbar
Pillow
psycopg2>=2.7,<3.0